Create package and move modules
