+++ src/imu_fusion/init_vectors.py
