+++ tests/test_mat_export.py
@@ -3,7 +3,7 @@ import pytest
 np = pytest.importorskip("numpy")
 pytest.importorskip("scipy")
 import scipy.io
-from utils import save_mat
+from imu_fusion.utils import save_mat
 
 def test_save_mat(tmp_path):
     data = {'a': np.array([1,2,3])}
