+++ tests/test_static_detection.py
@@ -1,8 +1,10 @@
 import os, sys
-sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
+repo_root = os.path.dirname(os.path.dirname(__file__))
+sys.path.insert(0, os.path.join(repo_root, "src"))
+sys.path.insert(0, repo_root)
 import pytest
 np = pytest.importorskip("numpy")
-from utils import detect_static_interval, is_static
+from imu_fusion.utils import detect_static_interval, is_static
 
 
 def test_detect_static_interval_basic():
