+++ tests/test_utils_additional.py
@@ -1,8 +1,10 @@
 import os, sys
-sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
+repo_root = os.path.dirname(os.path.dirname(__file__))
+sys.path.insert(0, os.path.join(repo_root, "src"))
+sys.path.insert(0, repo_root)
 import pytest
 np = pytest.importorskip("numpy")
-from utils import (
+from imu_fusion.utils import (
     adaptive_zupt_threshold,
     save_static_zupt_params,
     ecef_to_ned,
